requirements.txt
# Master QA Garden API Requirements
accelerate>=0.21.0
beautifulsoup4
beautifulsoup4>=4.12.0
browser-use
fastapi
fastapi>=0.104.0
fastapi>=0.104.1
fastapi>=0.115.0
google-generativeai
google-generativeai>=0.8.3
groq
groq>=0.4.0
groq>=0.4.2
httpx
httpx>=0.25.0
httpx>=0.27.0
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-openai
loguru>=0.7.2
playwright
playwright>=1.42.0
psutil
pydantic-settings>=2.6.0
pydantic>=2.10.0
pydantic>=2.5.0
pytest-html>=4.1.1
pytest-json-report>=1.5.0
pytest-playwright==0.4.4
pytest-rerunfailures>=13.0
pytest-xdist==3.5.0
pytest==8.0.0
pytest>=7.4.0
python-dotenv
python-dotenv==1.0.1
python-dotenv>=1.0.0
python-multipart>=0.0.12
requests
requests==2.31.0
requests>=2.31.0
tenacity
torch>=2.0.0
transformers>=4.35.2
uvicorn
uvicorn>=0.24.0
uvicorn[standard]>=0.32.0
websockets
websockets>=12.0
